moved create_instance fixture to global
