+++ cfme/tests/cloud/test_instance_power_control.py
@@ -43,24 +43,6 @@ def create_instance(appliance, provider, template_name):
     return instance
 
 
-@pytest.fixture(scope="function")
-def testing_instance(appliance, provider, small_template, setup_provider):
-    """ Fixture to provision instance on the provider
-    """
-    instance = create_instance(appliance, provider, small_template.name)
-    yield instance
-    instance.cleanup_on_provider()
-
-
-@pytest.fixture(scope="function")
-def testing_instance2(appliance, provider, small_template, setup_provider):
-    """ Fixture to provision instance on the provider
-    """
-    instance2 = create_instance(appliance, provider, small_template.name)
-    yield instance2
-    instance2.cleanup_on_provider()
-
-
 # This fixture must be named 'vm_name' because its tied to cfme/fixtures/virtual_machine
 @pytest.fixture(scope="function")
 def vm_name(testing_instance):
