+++ cfme/tests/cloud/test_tag_mapping.py
@@ -262,4 +262,3 @@ def test_ec2_tags(provider, request, ec2taggable, testing_instance):
     provider.refresh_provider_relationships()
     wait_for(provider.is_refreshed, func_kwargs={'refresh_delta': 10}, timeout=600)
     assert taggable.get_tag_value(tag_key) == tag_value
-
